+++ src/main/ruby/truffleruby/core/process.rb
@@ -618,9 +618,15 @@ module Process
   #       and called periodically.
   #
   def detach(pid)
+    pid = Truffle::Type.rb_convert_type(pid, Integer, :to_int)
     raise ArgumentError, 'Only positive pids may be detached' unless pid > 0
 
-    thread = Thread.new { Process.wait pid; $? }
+    thread = Thread.new do
+      Process.wait pid; $?
+    rescue Errno::ECHILD
+      # suppress exception "No child processes - No child process: <PID> (Errno::ECHILD)"
+    end
+
     thread[:pid] = pid
     def thread.pid; self[:pid] end