+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Compatibility:
 * Support accessing `dmark` and `dfree` fields for `RData` (#2771, @eregon).
 * Implement `rb_enc_nth()` (#2771, @eregon).
 * Support `offset` keyword argument for `String#unpack` and `String#unpack1` (@andrykonchin).
+* Fix `Process.detach` and cast `pid` argument to `Integer` (#2782, @andrykonchin).
 
 Performance: